Create An analysis of StackOverflow posts tags (DQ469 project))
This is an analysis of StackOverflow posts tags. The data we need can be downloaded from 
Stack Exchange Data Explorer (SEDE) (https://data.stackexchange.com/stackoverflow/query/new).
First, we'll analyze the frequency of using and viewing tags. We'll find the most popular. 
We ran a query against the SEDE DSSE database that the .csv - file for all the questions in 2019. Second, we'll analyze the dynamic of most popular tags. 
We'll use 'all_questions.csv' which presents data from 2014 year.